refactor: make getModule public
